configure checkstyle
