+++ checkstyle-suppressions.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0"?>
+<!DOCTYPE suppressions PUBLIC
+        "-//Checkstyle//DTD SuppressionFilter Configuration 1.0//EN"
+        "https://checkstyle.org/dtds/suppressions_1_0.dtd">
+<suppressions>
+    <suppress checks="MissingJavadocType" files="java$" />
+    <suppress checks="Indentation" files="java$" />
+    <suppress checks="CustomImportOrder" files="java$" />
+</suppressions>
