+++ src/main/java/io/jenkins/plugins/credentials/secretsmanager/AwsCredentialsProvider.java
@@ -1,0 +1,63 @@
+package io.jenkins.plugins.credentials.secretsmanager;
+
+import com.cloudbees.plugins.credentials.Credentials;
+import com.cloudbees.plugins.credentials.CredentialsProvider;
+import com.cloudbees.plugins.credentials.CredentialsStore;
+import com.cloudbees.plugins.credentials.common.IdCredentials;
+import edu.umd.cs.findbugs.annotations.NonNull;
+import hudson.Extension;
+import hudson.model.ItemGroup;
+import hudson.model.ModelObject;
+import hudson.security.ACL;
+import jenkins.model.Jenkins;
+import org.acegisecurity.Authentication;
+
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.List;
+import java.util.function.Supplier;
+import java.util.logging.Level;
+import java.util.logging.Logger;
+
+@Extension
+public class AwsCredentialsProvider extends CredentialsProvider {
+
+    private static final Logger LOG = Logger.getLogger(AwsCredentialsProvider.class.getName());
+
+    private final AwsCredentialsStore store = new AwsCredentialsStore(this);
+
+    private final Supplier<Collection<IdCredentials>> credentialsSupplier =
+            CredentialsSupplierFactory.create();
+
+    @Override
+    @NonNull
+    public <C extends Credentials> List<C> getCredentials(Class<C> type,
+                                                          ItemGroup itemGroup,
+                                                          Authentication authentication) {
+        if (ACL.SYSTEM.equals(authentication)) {
+            final ArrayList<C> list = new ArrayList<>();
+            for (IdCredentials credential : credentialsSupplier.get()) {
+                // is s a type of type then populate the list...
+                if (type.isAssignableFrom(credential.getClass())) {
+                    // cast to keep generics happy even though we are assignable..
+                    list.add(type.cast(credential));
+                }
+                LOG.log(Level.FINEST, "getCredentials {0} does not match", credential.getId());
+            }
+            return list;
+        }
+
+        return Collections.emptyList();
+    }
+
+    @Override
+    public CredentialsStore getStore(ModelObject object) {
+        return object == Jenkins.getInstance() ? store : null;
+    }
+
+    @Override
+    public String getIconClassName() {
+        return "icon-aws-secrets-manager-credentials-store";
+    }
+}
