+++ src/main/java/io/jenkins/plugins/credentials/secretsmanager/config/Filters.java
@@ -1,0 +1,50 @@
+package io.jenkins.plugins.credentials.secretsmanager.config;
+
+import io.jenkins.plugins.credentials.secretsmanager.Messages;
+
+import org.jenkinsci.Symbol;
+import org.kohsuke.stapler.DataBoundConstructor;
+import org.kohsuke.stapler.DataBoundSetter;
+
+import java.io.Serializable;
+
+import javax.annotation.Nonnull;
+
+import hudson.Extension;
+import hudson.model.AbstractDescribableImpl;
+import hudson.model.Descriptor;
+
+public class Filters extends AbstractDescribableImpl<Filters> implements Serializable {
+
+    /**
+     * Filter secrets received by their AWS tag. (To pass, a secret must have the specified tag key
+     * with one of the specified values.)
+     */
+    private Tag tag;
+
+    @DataBoundConstructor
+    public Filters(Tag tag) {
+        this.tag = tag;
+    }
+
+    public Tag getTag() {
+        return tag;
+    }
+
+    @DataBoundSetter
+    @SuppressWarnings("unused")
+    public void setTag(Tag tag) {
+        this.tag = tag;
+    }
+
+    @Extension
+    @Symbol("filters")
+    @SuppressWarnings("unused")
+    public static class DescriptorImpl extends Descriptor<Filters> {
+        @Override
+        @Nonnull
+        public String getDisplayName() {
+            return Messages.filters();
+        }
+    }
+}
