--- src/main/resources/io/jenkins/plugins/aws_secrets_manager_credentials_provider/Messages.properties
@@ -1,9 +1,0 @@
-awsSecretsManager = AWS Secrets Manager
-secretText = Secret text
-filters = Filters
-tag = Tag
-endpointConfiguration = Endpoint Configuration
-success = Success
-failure = Failure
-awsClientError = AWS client error
-awsServerError = AWS server error
