+++ src/main/resources/io/jenkins/plugins/credentials/secretsmanager/Messages.properties
@@ -1,0 +1,10 @@
+awsSecretsManager = AWS Secrets Manager
+secretText = Secret text
+filters = Filters
+tag = Tag
+endpointConfiguration = Endpoint Configuration
+success = Success
+failure = Failure
+awsClientError = AWS client error
+awsServerError = AWS server error
+couldNotRetrieveSecretError = Could not retrieve the secret for the credential with ID {0}. Please check that it exists in AWS Secrets Manager, and has not been soft-deleted.
