+++ src/main/resources/io/jenkins/plugins/credentials/secretsmanager/config/EndpointConfiguration/config_en.properties
