+++ src/main/resources/io/jenkins/plugins/credentials/secretsmanager/config/Filters/config.jelly
