+++ src/main/resources/io/jenkins/plugins/credentials/secretsmanager/config/PluginConfiguration/config_en.properties
