+++ src/test/java/io/jenkins/plugins/credentials/secretsmanager/config/AbstractPluginConfigurationTest.java
@@ -1,0 +1,15 @@
+package io.jenkins.plugins.credentials.secretsmanager.config;
+
+import org.junit.Test;
+
+public abstract class AbstractPluginConfigurationTest {
+
+    @Test
+    public abstract void shouldHaveDefaultConfiguration();
+
+    @Test
+    public abstract void shouldCustomiseEndpointConfiguration();
+
+    @Test
+    public abstract void shouldCustomiseTagFilter();
+}
